commit server changes

+++ server/Controllers/v1/AssignmentController.cs
@@ -18,6 +18,7 @@ namespace server.Controllers.v1
 
         [HttpGet]
         [ProducesResponseType(typeof(IEnumerable<AssignmentDTO>), StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
         [ProducesResponseType(StatusCodes.Status500InternalServerError)]
         public async Task<ActionResult> GetAll()
         {
@@ -29,9 +30,10 @@ namespace server.Controllers.v1
 
         [HttpGet("{id}")]
         [ProducesResponseType(typeof(AssignmentDTO), StatusCodes.Status200OK)]
-        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
-        [ProducesResponseType(StatusCodes.Status404NotFound)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
         public async Task<ActionResult> GetAssignment([FromRoute] Guid id)
         {
             var userId = authRepository.GetUserId();
@@ -46,17 +48,21 @@ namespace server.Controllers.v1
 
         [HttpDelete("{id}")]
         [ProducesResponseType(typeof(AssignmentDTO), StatusCodes.Status200OK)]
-        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
-        [ProducesResponseType(StatusCodes.Status404NotFound)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
+        [ProducesResponseType(StatusCodes.Status403Forbidden)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
         public async Task<ActionResult> DeleteAssignment([FromRoute] Guid id)
         {
             var userId = authRepository.GetUserId();
-            
+
             if (id.Equals("")) return BadRequest("Invalid ID sent.");
 
             Assignment? assignment = await repository.GetAssignmentById(id, userId);
             if (assignment == null) return NotFound("Assignment not found!");
+
+            if (!assignment.UserId.Equals(userId)) return Forbid();
 
             bool deleted = await repository.RemoveAssignment(assignment);
             if (!deleted) return BadRequest("Assignment not deleted.");
@@ -64,14 +70,34 @@ namespace server.Controllers.v1
             return Ok(mapper.Map<AssignmentDTO>(assignment));
         }
 
+        [HttpDelete]
+        [ProducesResponseType(typeof(string), StatusCodes.Status200OK)]
+        [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
+        [ProducesResponseType(StatusCodes.Status403Forbidden)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
+        public async Task<ActionResult> DeleteMyAssignments()
+        {
+            var userId = authRepository.GetUserId();
+
+            bool deleted = await repository.RemoveAllAssignments(userId);
+            if (!deleted) return BadRequest("Assignments not deleted.");
+
+            return Ok("Assignments successfully deleted.");
+        }
+
         [HttpPost]
         [ProducesResponseType(typeof(AssignmentDTO), StatusCodes.Status200OK)]
-        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
-        [ProducesResponseType(StatusCodes.Status404NotFound)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
         public async Task<ActionResult> CreateNewAssignment([FromBody] AssignmentCreateDTO createDTO)
         {
             var userId = authRepository.GetUserId();
+
+            if (!createDTO.UserId.Equals(userId)) return BadRequest();
 
             bool isExists = repository.AssignmentExists(createDTO.Title, userId);
             if (isExists) return BadRequest($"Assignment with name: {createDTO.Title} already exists.");
@@ -89,16 +115,18 @@ namespace server.Controllers.v1
 
         [HttpPut("{id}")]
         [ProducesResponseType(typeof(AssignmentDTO), StatusCodes.Status200OK)]
-        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
-        [ProducesResponseType(StatusCodes.Status404NotFound)]
         [ProducesResponseType(StatusCodes.Status400BadRequest)]
+        [ProducesResponseType(StatusCodes.Status401Unauthorized)]
+        [ProducesResponseType(StatusCodes.Status403Forbidden)]
+        [ProducesResponseType(StatusCodes.Status404NotFound)]
+        [ProducesResponseType(StatusCodes.Status500InternalServerError)]
         public async Task<ActionResult> UpdateExistingAssignment([FromRoute] Guid id, [FromBody] AssignmentUpdateDTO updateDTO)
         {
             var userId = authRepository.GetUserId();
 
-            if (id.Equals("") || updateDTO == null) return BadRequest();
-            if (!updateDTO.Id.Equals(id) || !updateDTO.UserId.Equals(userId)) return BadRequest();
-            
+            if (id.Equals("") || updateDTO == null || !updateDTO.Id.Equals(id)) return BadRequest();
+            if (!updateDTO.UserId.Equals(userId)) return Forbid();
+
             if (!ModelState.IsValid) return BadRequest(ModelState);
 
             bool isExists = repository.AssignmentExists(id, userId);
